Moved stuff to mol-util/object.ts
