+++ src/extensions/mvs/additions/selector.ts
@@ -13,14 +13,14 @@ import { MolScriptBuilder } from '../../../mol-script/language/builder';
 import { Expression } from '../../../mol-script/language/expression';
 import { UUID } from '../../../mol-util';
 import { arrayExtend, sortIfNeeded } from '../../../mol-util/array';
+import { mapArrayToObject, pickObjectKeys } from '../../../mol-util/object';
 import { ParamDefinition as PD } from '../../../mol-util/param-definition';
 import { capitalize } from '../../../mol-util/string';
-import { mapArrToObj, pickObjectKeys } from '../helpers/utils';
 import { MVSAnnotationStructureComponentParams, createMVSAnnotationStructureComponent } from './annotation-structure-component';
 
 
 /** Allowed values for a static selector */
-export const StaticSelectorChoice = new Choice(mapArrToObj(StaticStructureComponentTypes, t => capitalize(t)), 'all');
+export const StaticSelectorChoice = new Choice(mapArrayToObject(StaticStructureComponentTypes, t => capitalize(t)), 'all');
 export type StaticSelectorChoice = Choice.Values<typeof StaticSelectorChoice>
