+++ src/extensions/mvs/tree/molstar/conversion.ts
@@ -4,12 +4,12 @@
  * @author Adam Midlik <midlik@gmail.com>
  */
 
-import { omitObjectKeys, pickObjectKeys } from '../../helpers/utils';
 import { ConversionRules, addDefaults, condenseTree, convertTree, dfs } from '../generic/tree-utils';
 import { MolstarKind, MolstarNode, MolstarTree } from './molstar-tree';
 import { FullMVSTree, MVSTree, MVSTreeSchema } from '../mvs/mvs-tree';
 import { MVSDefaults } from '../mvs/mvs-defaults';
 import { MolstarParseFormatT, ParseFormatT } from '../mvs/param-types';
+import { omitObjectKeys, pickObjectKeys } from '../../../../mol-util/object';
 
 
 /** Convert `format` parameter of `parse` node in `MolstarTree`
